Session complète : refonte mobile, corrections critiques et préparation déploiement
Corrections front majeures :
- Bug sélection de dossier KYC figée sur le 1er élément (3 vues admin)
- Bug de virgule décimale française cassant les coordonnées GPS et le calcul
  de prix dans le JS (carte + réservation)
- Image de remplacement cassée (URL /static/static/ doublée) sur toute
  annonce sans photo
- Favoris factices, notes inventées, section "logements similaires" jamais
  peuplée, dropdown de notifications jamais connecté au backend

Refonte mobile de la fiche logement (galerie swipeable + vignettes, onglets
À propos/Galerie/Avis, barre de réservation fixe) et navigation mobile
(menu hamburger + panneau, bouton profil, barre du bas à 4 icônes).

Favoris en HTMX (plus de rechargement de page).

Performance : cache manquant sur PlatformSettings (relu sur chaque page).

Préparation déploiement Render : gunicorn respecte $PORT/$WEB_CONCURRENCY,
storage média Cloudinary optionnel (CLOUDINARY_URL), render.yaml (web +
Postgres gratuits).

+++ apps/reservations/services/services.py
@@ -1,5 +1,5 @@
+from decimal import Decimal
 from django.db import transaction
-from django.utils import timezone
 from apps.accounts.models import User
 from apps.properties.models import Property
 from apps.reservations.models import ReservationRequest, Reservation, ReservationStatusHistory, ReservationHistory
@@ -154,7 +154,7 @@ class ReservationService:
 
     @staticmethod
     @transaction.atomic
-    def confirm_payment(req: ReservationRequest, total_price: float) -> Reservation:
+    def confirm_payment(req: ReservationRequest, total_price: Decimal) -> Reservation:
         if req.status != ReservationStatusChoices.PAYMENT_PENDING:
             raise InvalidWorkflowTransition("La demande doit être au statut PENDING_PAYMENT.")
